+++ src/Method/UpdatingMessage/EditMessageCaption.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vjik\TelegramBot\Api\Method\UpdatingMessage;
+
+use Vjik\TelegramBot\Api\Request\HttpMethod;
+use Vjik\TelegramBot\Api\Request\TelegramRequestWithResultPreparingInterface;
+use Vjik\TelegramBot\Api\Type\InlineKeyboardMarkup;
+use Vjik\TelegramBot\Api\Type\Message;
+use Vjik\TelegramBot\Api\Type\MessageEntity;
+
+/**
+ * @see https://core.telegram.org/bots/api#editmessagecaption
+ */
+final readonly class EditMessageCaption implements TelegramRequestWithResultPreparingInterface
+{
+    /**
+     * @param MessageEntity[]|null $captionEntities
+     */
+    public function __construct(
+        private ?string $businessConnectionId = null,
+        private int|string|null $chatId = null,
+        private ?int $messageId = null,
+        private ?string $inlineMessageId = null,
+        private ?string $caption = null,
+        private ?string $parseMode = null,
+        private ?array $captionEntities = null,
+        private ?bool $showCaptionAboveMedia = null,
+        private ?InlineKeyboardMarkup $replyMarkup = null,
+    ) {
+    }
+
+    public function getHttpMethod(): HttpMethod
+    {
+        return HttpMethod::POST;
+    }
+
+    public function getApiMethod(): string
+    {
+        return 'editMessageCaption';
+    }
+
+    public function getData(): array
+    {
+        return array_filter(
+            [
+                'business_connection_id' => $this->businessConnectionId,
+                'chat_id' => $this->chatId,
+                'message_id' => $this->messageId,
+                'inline_message_id' => $this->inlineMessageId,
+                'caption' => $this->caption,
+                'parse_mode' => $this->parseMode,
+                'caption_entities' => $this->captionEntities === null ? null : array_map(
+                    static fn(MessageEntity $entity) => $entity->toRequestArray(),
+                    $this->captionEntities,
+                ),
+                'show_caption_above_media' => $this->showCaptionAboveMedia,
+                'reply_markup' => $this->replyMarkup?->toRequestArray(),
+            ],
+            static fn($value): bool => $value !== null,
+        );
+    }
+
+    public function prepareResult(mixed $result): Message|true
+    {
+        return $result === true
+            ? $result
+            : Message::fromTelegramResult($result);
+    }
+}