+++ tests/TelegramBotApiTest.php
@@ -431,6 +431,18 @@ final class TelegramBotApiTest extends TestCase
         $this->assertTrue($result);
     }
 
+    public function testEditMessageCaption(): void
+    {
+        $api = $this->createApi([
+            'ok' => true,
+            'result' => true,
+        ]);
+
+        $result = $api->editMessageCaption();
+
+        $this->assertTrue($result);
+    }
+
     public function testEditMessageText(): void
     {
         $api = $this->createApi([