+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - New #70: Add `getBusinessConnection` method.
 - New #71: Add `getMyDefaultAdministratorRights` and `setMyDefaultAdministratorRights` methods.
 - New #72: Add `editMessageText` method and `ParseMode` class with constants.
+- New #73: Add `editMessageCaption` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.